context.py: Fix duplicate step names across tests
I have fixed an issue where the step names must be unique
throughout the entire test suite.  If, for example, there were
two tests that had defined a `do_foo` step it would cause the
second definition of `do_foo` to throw a `RuntimeError` instead
of running even though the function is in an entirely different file.

To get around this we now will destroy the global step collection
at the setup of each test class.  This means that can only be one
test class per file However a simple workaround would be to create
a base class and have any other test classes in your module inherit
from it.